% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twinflex.R
\name{twinflex}
\alias{twinflex}
\title{Estimates different types of twin models}
\usage{
twinflex(
  acevars = NULL,
  zyg = "zyg",
  sep = "",
  data = NULL,
  covvars = NULL,
  covariance = FALSE,
  type = "Chol",
  moderation = NULL,
  ModCov = "DV",
  ordinal = NULL,
  TryHard = FALSE,
  Tries = 10,
  exh = TRUE,
  Optimizer = "SLSQP",
  lboundACE = TRUE,
  dzA = 0.5,
  dzC = 1
)
}
\arguments{
\item{acevars}{a string vector containing the variables whose variance/covariance you want to decompose. For multivariate models, just use several variables. At the moment only the Cholesky parametrization is implemented. The first variable in the acevars vector is considered to be the first in this causal or temporary sense as well; the second variable as the second and so on. So have in mind to check the order of your variables in the multivariate case.}

\item{zyg}{please recode your zygosity variable as follows: 1 = MZ and 2 = DZ, the variable name must be \code{zyg}}

\item{sep}{a separator referring to the element that separates the variable names and the twin specific suffix. For example, the variables in the data frame are called \code{IQ_1} for twin 1 and \code{IQ_2} for twin 2. The \code{acevars} argument then is \code{acevars = 'IQ'} and the separator is \code{separator = '_'}}

\item{data}{a raw data frame (please just remove any SPSS or Stata labels/notes etc. before).}

\item{covvars}{a string vector with covariates. It is possible to use long-formatted (1 variable per twin pair) or wide-formatted (1 variable per twin) variables. Please make sure that you use the wide format only for variables that have a within-twin-pair variance.}

\item{covariance}{a logical to define whether the covariates should be modelled as random covariates in the model for the covariance matrix or as fixed covariates in the model for the mean.}

\item{type}{a string vector to define the "type" of a multivariate model. For a Cholesky model (default) use \code{type = 'Chol'}. For the "ACE-beta" model, use \code{type = 'aceb'}}

\item{moderation}{a string vector indicating a moderated path coefficient. Notation is comparable to MPLUS. The effects of covariates cannot be moderated. For a univariate twin model the argument \verb{moderation = Y BY Mod} with \code{Y} as the outcome and \code{Mod} as the moderator refers to a moderation of the effects of the ACE components on \code{Y} by \code{Mod}. For a multivariate model there are two options which can be combined: a) the argument \verb{moderation = Y BY Mod} refers to a moderation of the effects of the ACE components unique to \code{Y} on \code{Y} by \code{Mod}; b) the argument \verb{moderation = X -> Y BY Mod} refers to a moderation of the effects of the ACE components of \code{X} on \code{Y} (and the phenotypic effect of \code{X} on \code{Y} when \code{type = 'aceb'}) by \code{Mod}.}

\item{ModCov}{a string vector indicating whether main effect of moderator should be estimated for all variables (\code{ModCov = "All"}) or for only the variables for which the incoming paths are moderated (\code{ModCov = "DV"}). Default is \code{ModCov = "DV"}.}

\item{ordinal}{a string vector defining the ordinal variables (binary or ordinal)}

\item{TryHard}{a logical to define whether TryHard should be used (see OpenMx documentation for details). Recommended for complex models.}

\item{Tries}{a numeric indicating the number of attempts to run the model in addition to the first (see OpenMx documentation for details). Default is 10.}

\item{exh}{a logical indicating model fitting stops making additional attempts once it reaches an acceptable solution (when \code{exh = FALSE} as the default).}

\item{Optimizer}{a string vector indicating the optimizer. Default is \code{Optimizer = 'SLSQP'} (see OpenMx documentation for details).}

\item{lboundACE}{a logical indicating whether to set lower bounds for the estimation of the univariate ACE effects.}

\item{dzA}{a numerical vector indicating the genetic correlation for DZ twins (changeable to adjust for assortative mating). Default is \code{dzA = 0.5}}

\item{dzC}{a numerical vector indicating the shared environmental correlation for DZ twins. Default is \code{dzC = 1}. Change to \code{dzC = 0.25} to estimate an ADE model.}
}
\description{
twinflex is a wrapper function for OpenMx. It allows you to estimate complex twin models with just one line of code. Furthermore you can combine different types of twin models (univariate, GxE, with or without covariates, Cholesky vs. ACE-beta) individually.
}
